+++ ohos_luaarkts/CHANGELOG.md
@@ -1,6 +1,7 @@
 **CHANGELOG.md**
 ## 1.0.2
 - Optimized Shared Library Compilation: O3 and LTO Support
+- Fix the return value issue of T2lCallFunction
 
 ## 1.0.2-rc.1
 - Replace luaJIT with the Lua library